requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.4
aiosignal==1.3.1
albumentations==1.4.3
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
anykeystore==0.2
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
better-exceptions==0.3.3
bleach==6.1.0
bypy==1.8.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
crcmod==1.7
cryptacular==1.6.2
cryptography==42.0.5
cycler==0.12.1
DateTime==5.5
debugpy
decorator
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
dlib==19.24.4
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.13
efficientnet-pytorch==0.7.1
einops==0.7.0
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.26.0
gradio_client==0.15.1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hupper==1.12.1
idna==3.7
ilearnplus==1.5
imageio==2.34.0
importlib_resources==6.4.0
imutils==0.5.4
intel-openmp==2024.1.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.4
lightgbm==4.3.0
lightning-utilities==0.11.2
lmdb==1.4.1
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2024.1.0
ml-collections==0.1.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio
networkx==3.3
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openTSNE==1.0.1
openxlab==0.0.38
orjson==3.10.1
oss2==2.17.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso
PasteDeploy==3.1.0
pbkdf2==1.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs
pretrainedmodels==0.7.4
prometheus_client==0.20.0
prompt-toolkit
protobuf==4.25.3
psutil==5.9.8
ptyprocess
pure-eval
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil
python-json-logger==2.0.7
python-multipart==0.0.9
python3-openid==3.2.0
pytorch-lightning==2.2.1
pytorch-metric-learning==2.5.0
pytorch-ranger==0.1.1
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.0.0
QDarkStyle==3.2.3
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.18.0
ruff==0.3.7
safetensors==0.4.2
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==1.45.0
setproctitle==1.3.3
shellingham==1.5.4
sip==6.8.3
six
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
terminado==0.18.1
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.2
tinycss2==1.2.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torch-optimizer==0.3.0
torchcam==0.4.0
torchmetrics==1.3.2
torchstat==0.0.7
torchvision==0.17.2
tornado
tqdm==4.65.2
traitlets
transaction==4.0
translationstring==1.4
triton==2.2.0
ttach==0.0.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.29.0
velruse==1.1.1
venusian==3.1.0
wandb==0.16.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xgboost==2.0.3
yarg==0.1.9
yarl==1.9.4
zope.deprecation==5.0
zope.interface==6.4.post2
zope.sqlalchemy==3.1
